% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color}
\alias{color}
\title{Color list}
\format{
A list with 3 elements:
\describe{
 \item{color1}{Color for gender breakdown.}
 \item{color2}{Color for the language insights.}
 \item{color3}{Color for trend analysis.}
 }
}
\source{
Generated for this Shiny app.
}
\usage{
color
}
\description{
A list of colors used in the app.
}
\keyword{datasets}
